refactor sentence parser
