+++ src/1-one/tokenize/methods/01-sentences/01-simple-split.js
