+++ src/1-one/tokenize/methods/01-sentences/is-sentence.js
